Move citation parser to its own file
- Start writing a single unit test
